pyproject.toml
[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deep-research-newsletter"
version = "0.1.0"
description = "Agentic deep-research pipeline for turning a paper into a structured newsletter."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "FlagEmbedding>=1.3.5",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.5.4",
    "openai>=1.54.0",
    "pdf2image>=1.17.0",
    "Pillow>=10.4.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "rich>=14.0.0",
    "tavily-python>=0.5.0",
    "tiktoken>=0.8.0",
    "torch>=2.4.0",
    "typing-extensions>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires local documents or heavier runtime dependencies",
]
